% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mni_dados_basicos_xml.R
\name{mni_dados_basicos_xml}
\alias{mni_dados_basicos_xml}
\title{Lê dados do xml}
\usage{
mni_dados_basicos_xml(xml)
}
\arguments{
\item{xml}{Objeto xml lido com xml2::read_xml.}
}
\value{
tibble
}
\description{
Lê dados do xml
}
