% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mni_dataset_figshare_link.R
\docType{data}
\name{mni_dataset_figshare_link}
\alias{mni_dataset_figshare_link}
\title{Links from Figshare to MNI data}
\format{A \code{data.frame} with columns
\describe{
\item{url}{download url MNI}
\item{name}{filename to download}
}}
\usage{
mni_dataset_figshare_link
}
\description{
Links from Figshare to MNI data
}
\keyword{datasets}
