% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mni_data.R
\name{mni_datasets}
\alias{mni_datasets}
\title{MNI Datasets}
\usage{
mni_datasets(type = c("nifti", "minc1", "minc2"))
}
\arguments{
\item{type}{Type of imaging, either MINC or NIfTI types}
}
\value{
A \code{data.frame} with names and groups
}
\description{
MNI Datasets
}
\examples{
mni_datasets()
}
