% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mni_figshare.R
\docType{data}
\name{mni_figshare}
\alias{mni_figshare}
\title{Links to each template on Figshare}
\format{A \code{data.frame}
\describe{
\item{download_url}{download url from Figshare}
\item{name}{filename to download}
\item{id}{Figshare identifier of the file}
\item{size}{size of the file if downloaded}
}}
\usage{
mni_figshare
}
\description{
Links to each template on Figshare
}
\keyword{datasets}
