% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mni_dataset_figshare_link.R
\docType{data}
\name{mni_dataset_figshare_link}
\alias{mni_dataset_figshare_link}
\title{Links from Figshare to MNI data}
\format{
A \code{data.frame} with columns
\describe{
\item{url}{download url MNI}
\item{name}{filename to download}
}
}
\usage{
mni_dataset_figshare_link
}
\description{
Links from Figshare to MNI data
}
\examples{
df = mni_datasets()
dim(df)
\dontrun{
df = dplyr::left_join(df, mni_dataset_figshare_link, by = "url")
head(df)
dim(df)
df = dplyr::left_join(df, mni_figshare, by = "name")
head(df)
dim(df)
}

}
\keyword{datasets}
